replacing layers file with module
