+++ nvtabular/framework_utils/tensorflow/layers/__init__.py
@@ -1,0 +1,2 @@
+from embedding import ScalarDenseFeatures, ScalarWideFeatures
+from interaction import DotProductInteraction
